Azure/enterprise-arch-agent/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-dotenv==1.1.1
pydantic==2.11.7
azure-ai-projects==1.0.0b12
azure-identity==1.24.0
openai==1.101.0
